requirements.txt
pandas
camelot-py[cv]
openpyxl
requests
folium
